Updated readme to use the correct commands
The previous version of the README listed PowerShell commands that do not work as intended. For example using "$env:HOME" in quotes as occurs on line 129 of the previous README will treat that as a string literal, and add the exact string "$env:HOME" to your PATH. or otherwise will drop it entirely, leaving you with "\.pyenv.." in your path. Using string interpolation within the SetEnvironmentVariables method is preferred.

Also, it does not appear that there is a _need_ for the default install to use SYSTEM level environment variables. All pyenv operations are launched in the user context and thus no system variables are needed. If this is not correct, then the same method should be used for adding the SYSTEM variable, simply with the third argument should be set to "Machine" instead of "User"

Finally, added a more complete workaround for the Python/App Installer issue. Though the suggested changes made to the PowerShell commands should insert the as a higher order in PATH, removing the alias altogether is more effective.