Add failing test for exec.
The reason it fails is that `pyenv.bat exec` gets the list of active versions from `pyenv.bat vname` and then starts iterating over it. For every line the vname gets appended to the versions folder and then normalized into %bindir%. :exec of pyenv.bat later checks to see if %bindir% is set or not so it can show an error if it is not.

That should have been the last use of %bindir%, but is is not. The immidiate reason it does this is that any shim taken from the Scripts folder runs `pyenv.bat exec Scripts\<the original command goes here>`. `Scripts\something` is a path so Windows does not search the path for something to run and instead looks in the current folder and more often then not won't find anything and throw an error.
The underlying cause is that is that the shims for files located in Scripts are prepending `Scripts\`.

The next commit is going to address that.