fix Chocolatey installation link in README
The link in the *Installation* section to the Chocolatey instructions was broken in the rendered HTML.

For GitHub Pages the on-page anchor links need to be lowercase. This isn't well documented but when testing in both chromium and firefox browsers, changing the inner-link from (#Chocolatey) -> (#chocolatey) fixed the broken link.