Separate out bare shims directory
PowerShell prefers the file `python3.bat` to `python3` when these two
files share a directory on the `PATH`, which we want, but it also
prefers `python3.11` to `python3.11.bat` when these two files share a
directory on the `PATH`, which we don't want. Separating out shims
without extensions to a `$pyenv_home\bare_shims` directory that's
after `$pyenv_home\shims` on the `PATH` resolves this issue.